requirements.txt
asgiref==3.8.0
Django==5.0.3
django-filter==24.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==24.3.0
inflection==0.5.1
packaging==24.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1